manage taxes and dahsboard changes

+++ HrMaxx.OnlinePayroll.Models/DataModel/FITTaxTable.cs
@@ -17,11 +17,11 @@ namespace HrMaxx.OnlinePayroll.Models.DataModel
         public int Id { get; set; }
         public Nullable<int> PayrollPeriodID { get; set; }
         public string FilingStatus { get; set; }
-        public Nullable<double> StartRange { get; set; }
-        public Nullable<double> EndRange { get; set; }
-        public Nullable<double> FlatRate { get; set; }
-        public Nullable<double> AdditionalPercentage { get; set; }
-        public Nullable<double> ExcessOvrAmt { get; set; }
+				public Nullable<decimal> StartRange { get; set; }
+				public Nullable<decimal> EndRange { get; set; }
+				public Nullable<decimal> FlatRate { get; set; }
+				public Nullable<decimal> AdditionalPercentage { get; set; }
+				public Nullable<decimal> ExcessOvrAmt { get; set; }
         public Nullable<int> Year { get; set; }
     }
 }
